% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_fncs.R
\name{calc_roc}
\alias{calc_roc}
\title{Calculate ROC curve metrics (two-class case)}
\usage{
calc_roc(pred_vec, labels, thresholds = seq(0.01, 1, 0.01))
}
\arguments{
\item{pred_vec}{A numeric vector of predicted probabilities}

\item{labels}{A factor vector of class labels (must use \code{TRUE} and \code{FALSE})}

\item{threshholds}{a numeric of threshholds to test.}
}
\value{
a three column matrix with the true positive rate (tpr), false discovery,
rate (fdr), and true negative rate (tnr) associated with each threshold value.
}
\description{
Calcluates the trade-off between true positive rate (tpr) and 
false discovery rate (fdr), true negative rate (tnr), and overall accuracy 
for a set of threshold probabilities.
}

