% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_fncs.R
\name{class_logloss}
\alias{class_logloss}
\title{Calculate the two-class logloss}
\usage{
class_logloss(p_vec, class_vec, w_vec = NULL)
}
\arguments{
\item{p_vec}{a vector of predicted probabilities for the positive class}

\item{class_vec}{A vector of class labels}

\item{w_vec}{An optional vector of observation weights}
}
\description{
Calculates the logloss defined:
\eqn{\frac{1}{N}\sum_i^N y_{i} log(p_i)} similar to \code{\link{multiclass_logloss}} 
but for a two-class problem
}

