% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitting.R
\name{create_partitions}
\alias{create_partitions}
\title{Create data partitions}
\usage{
create_partitions(df, dep_var, level = NULL, n = 100L, major_class_wt = 1,
  seed = 379L, test_pct = 0.33, binomial = TRUE)
}
\arguments{
\item{df}{A \code{data.frame} for analysis}

\item{dep_var}{A character string denoting the dependent variable in \code{df}.}

\item{level}{level of interest. If \code{NULL} takes the 2nd level of a factor variable
or the 2nd unique value from a non-factor variable.}

\item{n}{An integer denoting the number of ensembles to build. Defaults to \code{100L}.}

\item{major_class_wt}{Controls the number of major class cases selected in each
partition as a multiple of the number of minority class observations. Defaults to \code{1}, 
which will produce equal sized sets of minority and non-minority class in each partition.
Must be greater than or equal to 1.}

\item{seed}{An integer. Seed for reproducibility. Defaults to \code{379L}.}

\item{test_pct}{A number in (0,1) specifying the size of the test dataset as a percentage.
Defaults to \code{0.33}}

\item{binomial}{Logical. Does the response variable follow a Binomial distribution? Defaults 
to \code{TRUE}.}
}
\description{
Create n data partitions for training with equal sized classes via resampling.
}

