% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_ensemble}
\alias{glm_ensemble}
\title{Fit a GLM Ensemble model}
\usage{
glm_ensemble(df, dep_var, cols = which(names(df) != dep_var), n = 100L,
  level = NULL, major_class_wt = 1, seed = 379L, test_pct = 0.33,
  direction = "backward", family = binomial(link = "logit"),
  leave_cores = NULL)
}
\arguments{
\item{df}{A \code{data.frame} for analysis}

\item{dep_var}{A character string denoting the dependent variable in \code{df}.}

\item{cols}{A vector of column indices corresponding the the variables you wish to regress on.
This allows for variable (de)selection prior to model building. Defaults to using all columns.}

\item{n}{An integer denoting the number of ensembles to build; defaults to \code{100L}.}

\item{level}{level of interest. If \code{NULL} takes the 2nd level of a factor variable
or the 2nd unique value from a non-factor variable.}

\item{major_class_wt}{Controls the number of major class cases selected in each
partition as a multiple of the number of minority class observations. Defaults to \code{1}, 
which will produce equal sized sets of minority and non-minority class in each partition.
Must be greater than or equal to 1.}

\item{seed}{An integer. Seed for reproducibility; defaults to \code{379L}.}

\item{test_pct}{A number in (0,1) specifying the size of the test dataset as a percentage.}

\item{direction}{A character vector for the step process.}

\item{family}{Used to specify the details of the glm methods. See \code{\link{[stats]family}}}

\item{leave_cores}{An integer for number of cores to leave unused.}
}
\value{
A list of with a matrix of coefficients from each ensemble element, the 
element weights, and the weighted coefficient estimates.
}
\description{
Fit a GLM Ensemble model via \code{\link{[stats]glm}} in parallel. Training datasets
are chosen to have equal number of each class; and, a single dataset is used to determine
prediction error (and ensemble weight) for each element in the ensemble. Each GLM does variable
selection via the \code{\link{[stats]step}} function with non-verbose output.
}

