% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_fncs.R
\name{multiclass_logloss}
\alias{multiclass_logloss}
\title{Calculate the multiclass logloss}
\usage{
multiclass_logloss(p_mat, class_mat, w_vec = NULL)
}
\arguments{
\item{p_mat}{An N x m matrix of predicted probabilities for a multiclass
problem. N == number of observations. m = number of classes. 
It is assumed that the row-sums all == 1}

\item{class_mat}{An N x m {0,1} integer matrix giving the true classes
of the observations. It is assumed that the row-sums all == 1.}

\item{w_vec}{An optional vector of observation weights. Defaults to NULL 
ie - equal weights}
}
\value{
A numeric scalar
}
\description{
Calculates the multiclass logloss defined:
\eqn{\frac{1}{N}\sum_i^N\sum_j^M y_{ij} log(p_ij)}
}

