% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{print.glmEnsemble}
\alias{print.glmEnsemble}
\title{Printing GLM-Ensemble fits}
\usage{
\method{print}{glmEnsemble}(object, ...)
}
\arguments{
\item{object}{An object of class "glmEnsemble".}

\item{...}{further arguments passed to or from other methods}
}
\description{
Generic methods for class "glmEnsemble"
}

