% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{tt_glm_par}
\alias{tt_glm_par}
\title{Train-Test a GLM ensemble in parallel}
\usage{
tt_glm_par(train_list, test_dat, out_vec, direction, family = binomial(link =
  "logit"), leave_cores = 1)
}
\arguments{
\item{train_list}{A \code{list} of training \code{data.frames} for ensemble training}

\item{test_dat}{A \code{data.frame} for calculating error rates from ensemble elements.}

\item{out_vec}{A named \code{vector} "backbone" on which to place ensemble coefficients.}

\item{direction}{A character string for the direction used in \code{\link{[stats]step}}}

\item{family}{Used to specify the details of the glm methods. See \code{\link{[stats]family}}}

\item{leave_cores}{An integer for number of cores to leave unused. Defaults to \code{1L}.}
}
\description{
Takes a list of training datasets and a single test dataset and builds 
an ensemble of GLMs in parallel. Additionally calculates ensemble weights based
on element accuracy.
}

